+++ packages/contracts/contracts/core/component/dao-authorizable/DaoAuthorizableCloneable.sol
@@ -2,10 +2,10 @@
 
 pragma solidity 0.8.10;
 
-import "@openzeppelin/contracts/proxy/utils/Initializable.sol";
+import {Initializable} from "@openzeppelin/contracts/proxy/utils/Initializable.sol";
 
-import {IDAO} from "../IDAO.sol";
-import {DaoAuthorizableBase} from "./DaoAuthorizableBase.sol";
+import {IDAO} from "../../IDAO.sol";
+import {DaoAuthorizableBase} from "./bases/DaoAuthorizableBase.sol";
 
 /// @title DaoAuthorizableCloneable
 /// @author Aragon Association - 2022