+++ packages/contracts/contracts/plugin/PluginRepo.sol
@@ -11,6 +11,7 @@ import {Address} from "@openzeppelin/contracts/utils/Address.sol";
 import {PermissionManager} from "../core/permission/PermissionManager.sol";
 import {_uncheckedIncrement} from "../utils/UncheckedMath.sol";
 import {PluginSetup} from "./PluginSetup.sol";
+import {IPluginSetup} from './PluginSetup.sol';
 import {IPluginRepo} from "./IPluginRepo.sol";
 
 /// @title PluginRepo
@@ -18,9 +19,9 @@ import {IPluginRepo} from "./IPluginRepo.sol";
 /// @notice The plugin repository contract required for managing and publishing different plugin versions within the Aragon DAO framework following the [Semantic Versioning 2.0.0](https://semver.org/) convention.
 //TODO Rename to PluginSetupRepo?
 contract PluginRepo is
-    IPluginRepo,
     Initializable,
     ERC165Upgradeable,
+    IPluginRepo,
     UUPSUpgradeable,
     PermissionManager
 {
@@ -102,7 +103,7 @@ contract PluginRepo is
         // the below approach aims to still return custom error which not possible with try/catch..
         // NOTE: also checks if _pluginSetup is a contract and reverts if not.
         bytes memory data = _pluginSetup.functionCall(
-            abi.encodeWithSelector(ERC165.supportsInterface.selector, type(PluginSetup).interfaceId)
+            abi.encodeWithSelector(ERC165.supportsInterface.selector, type(IPluginSetup).interfaceId)
         );
 
         // NOTE: if data contains 32 bytes that can't be decoded with uint256
@@ -253,8 +254,13 @@ contract PluginRepo is
         auth(address(this), UPGRADE_PERMISSION_ID)
     {}
 
-    /// @inheritdoc ERC165Upgradeable
+    /// @notice Checks if this or the parent contract supports an interface by its ID.
+    /// @param interfaceId The ID of the interace.
+    /// @return bool Returns true if the interface is supported.
     function supportsInterface(bytes4 interfaceId) public view virtual override returns (bool) {
-        return interfaceId == type(IPluginRepo).interfaceId || super.supportsInterface(interfaceId);
+        return
+            interfaceId == type(IPluginRepo).interfaceId ||
+            interfaceId == type(UUPSUpgradeable).interfaceId ||
+            super.supportsInterface(interfaceId);
     }
 }