+++ packages/contracts/test/core/permission/permission-manager.ts
@@ -583,8 +583,7 @@ describe.only('Core: PermissionManager', function () {
       }
     });
 
-    // TODO: find out what is wrong with this test
-    it.skip('should revert on errors', async () => {
+    it('throws Unauthorized error when caller does not have ROOT_PERMISSION_ID permission', async () => {
       const signers = await ethers.getSigners();
       const bulkItems: ItemMultiTarget[] = [
         {
@@ -596,13 +595,15 @@ describe.only('Core: PermissionManager', function () {
         },
       ];
 
-      await expect(pm.bulkOnMultiTarget(bulkItems)).to.be.revertedWith(
+      await expect(
+        pm.connect(signers[2]).bulkOnMultiTarget(bulkItems)
+      ).to.be.revertedWith(
         customError(
           'Unauthorized',
           pm.address,
           signers[1].address,
-          signers[0].address,
-          ADMIN_PERMISSION_ID
+          signers[2].address,
+          ROOT_PERMISSION_ID
         )
       );
     });