+++ packages/contracts/test/plugin/plugin-setup-processor.ts
@@ -1,20 +1,31 @@
 import {expect} from 'chai';
 import console from 'console';
 import {ethers} from 'hardhat';
+import {SignerWithAddress} from '@nomiclabs/hardhat-ethers/signers';
+
 import {
   PluginSetupProcessor,
+  PluginCloneableMock,
   PluginSetupV1Mock,
   PluginSetupV2Mock,
   PluginSetupV1MockBad,
-  PluginRepoFactory,
   PluginRepoRegistry,
   PluginRepo,
 } from '../../typechain';
+
 import {customError} from '../test-utils/custom-error-helper';
+import {deployENSSubdomainRegistrar} from '../test-utils/ens';
 
 import {deployNewDAO} from '../test-utils/dao';
-import {decodeEvent} from '../test-utils/event';
-import {prepareInstallation} from '../test-utils/plugin-setup-processor';
+import {findEvent} from '../test-utils/event';
+import {
+  deployPluginSetupProcessor,
+  prepareInstallation,
+} from '../test-utils/plugin-setup-processor';
+import {
+  deployPluginRepoFactory,
+  deployPluginRepoRegistry,
+} from '../test-utils/repo';
 
 enum Op {
   Grant,
@@ -38,7 +49,6 @@ const EVENTS = {
 const EMPTY_DATA = '0x';
 
 const AddressZero = ethers.constants.AddressZero;
-// default oracle address emitted from permission manager
 const ADDRESS_TWO = `0x${'00'.repeat(19)}02`;
 
 const EMPTY_ID = `0x${'00'.repeat(32)}`;
@@ -54,43 +64,22 @@ const APPLY_UNINSTALLATION_PERMISSION_ID = ethers.utils.id(
 const SET_REPO_REGISTRY_PERMISSION_ID = ethers.utils.id(
   'SET_REPO_REGISTRY_PERMISSION'
 );
-const PLUGIN_REGISTER_PERMISSION_ID = ethers.utils.id(
-  'PLUGIN_REGISTER_PERMISSION'
+const REGISTER_PLUGIN_REPO_PERMISSION_ID = ethers.utils.id(
+  'REGISTER_PLUGIN_REPO_PERMISSION'
 );
 const UPGRADE_PERMISSION_ID = ethers.utils.id('UPGRADE_PERMISSION');
 
-// Util Helper Functions
-async function getPluginRepoFactoryMergedABI() {
-  // @ts-ignore
-  const PluginRepoRegistryArtifact = await hre.artifacts.readArtifact(
-    'PluginRepoRegistry'
-  );
-  // @ts-ignore
-  const PluginRepoFactoryArtifact = await hre.artifacts.readArtifact(
-    'PluginRepoFactory'
-  );
+const REGISTER_ENS_SUBDOMAIN_PERMISSION_ID = ethers.utils.id(
+  'REGISTER_ENS_SUBDOMAIN_PERMISSION'
+);
 
-  const _merged = [
-    ...PluginRepoFactoryArtifact.abi,
-    ...PluginRepoRegistryArtifact.abi.filter((f: any) => f.type === 'event'),
-  ];
-
-  // remove duplicated events
-  const merged = _merged.filter(
-    (value, index, self) =>
-      index === self.findIndex(event => event.name === value.name)
-  );
-
-  return {
-    abi: merged,
-    bytecode: PluginRepoFactoryArtifact.bytecode,
-  };
-}
+let counter = 0;
 
 describe('Plugin Setup Processor', function () {
-  let signers: any;
+  let signers: SignerWithAddress[];
   let psp: PluginSetupProcessor;
   let pluginRepo: PluginRepo;
+  let pluginCloneableMock: PluginCloneableMock;
   let pluginSetupV1Mock: PluginSetupV1Mock;
   let pluginSetupMockRepoAddress: any;
   let pluginSetupV2Mock: PluginSetupV2Mock;
@@ -107,6 +96,12 @@ describe('Plugin Setup Processor', function () {
   });
 
   before(async () => {
+    // Directly deploy PluginCloneableMock
+    const _PluginCloneableMock = await ethers.getContractFactory(
+      'PluginCloneableMock'
+    );
+    pluginCloneableMock = await _PluginCloneableMock.deploy();
+
     // PluginSetupV1
     const PluginSetupV1Mock = await ethers.getContractFactory(
       'PluginSetupV1Mock'
@@ -122,40 +117,41 @@ describe('Plugin Setup Processor', function () {
     // Managing DAO that have permission to manage PluginSetupProcessor
     managingDao = await deployNewDAO(ownerAddress);
 
-    // PluginRepoRegistry
-    const PluginRepoRegistry = await ethers.getContractFactory(
-      'PluginRepoRegistry'
-    );
-    pluginRepoRegistry = await PluginRepoRegistry.deploy();
-    await pluginRepoRegistry.initialize(managingDao.address);
-
-    // PluginRepoFactory
-    const {abi, bytecode} = await getPluginRepoFactoryMergedABI();
-    const PluginRepoFactory = new ethers.ContractFactory(
-      abi,
-      bytecode,
-      signers[0]
+    // ENS subdomain Registry
+    const ensSubdomainRegistrar = await deployENSSubdomainRegistrar(
+      signers[0],
+      managingDao,
+      'dao.eth'
     );
 
-    pluginRepoFactory = await PluginRepoFactory.deploy(
-      pluginRepoRegistry.address
+    // Plugin Repo Registry
+    pluginRepoRegistry = await deployPluginRepoRegistry(
+      managingDao,
+      ensSubdomainRegistrar
+    );
+
+    // Plugin Repo Factory
+    pluginRepoFactory = await deployPluginRepoFactory(
+      signers,
+      pluginRepoRegistry
     );
 
     // Grant `PLUGIN_REGISTER_PERMISSION` to `PluginRepoFactory`.
     await managingDao.grant(
       pluginRepoRegistry.address,
       pluginRepoFactory.address,
-      PLUGIN_REGISTER_PERMISSION_ID
+      REGISTER_PLUGIN_REPO_PERMISSION_ID
     );
 
-    // PluginSetupProcessor
-    const PluginSetupProcessor = await ethers.getContractFactory(
-      'PluginSetupProcessor'
+    // Grant `REGISTER_ENS_SUBDOMAIN_PERMISSION` to `PluginRepoFactory`.
+    await managingDao.grant(
+      ensSubdomainRegistrar.address,
+      pluginRepoRegistry.address,
+      REGISTER_ENS_SUBDOMAIN_PERMISSION_ID
     );
-    psp = await PluginSetupProcessor.deploy(
-      managingDao.address,
-      pluginRepoRegistry.address
-    );
+
+    // Plugin Setup Processor
+    psp = await deployPluginSetupProcessor(managingDao, pluginRepoRegistry);
   });
 
   beforeEach(async function () {
@@ -164,14 +160,16 @@ describe('Plugin Setup Processor', function () {
 
     // Create and register a plugin on the PluginRepoRegistry
     const tx = await pluginRepoFactory.createPluginRepoWithVersion(
-      'PluginSetupV1Mock',
+      `PluginSetupV1Mock-${counter}`,
       [1, 0, 0],
       pluginSetupV1Mock.address,
       '0x00',
       ownerAddress
     );
 
-    const event = await decodeEvent(tx, EVENTS.PluginRepoRegistered);
+    counter++;
+
+    const event = await findEvent(tx, EVENTS.PluginRepoRegistered);
 
     pluginSetupMockRepoAddress = event.args.pluginRepo;
 
@@ -212,7 +210,7 @@ describe('Plugin Setup Processor', function () {
     });
 
     describe('PrepareInstallation', function () {
-      it('Reverts if `PluginSetupRepo` do not exist on `PluginRepoRegistry`', async () => {
+      it('reverts if `PluginSetupRepo` do not exist on `PluginRepoRegistry`', async () => {
         const data = '0x';
         const pluginSetupRepoAddr = ADDRESS_TWO;
 
@@ -226,7 +224,7 @@ describe('Plugin Setup Processor', function () {
         ).to.be.revertedWith(customError('EmptyPluginRepo'));
       });
 
-      it('Reverts if installation already prepared', async () => {
+      it('reverts if installation already prepared', async () => {
         const pluginSetupBad = pluginSetupV1MockBad.address;
 
         const data1 = ethers.utils.defaultAbiCoder.encode(
@@ -256,10 +254,9 @@ describe('Plugin Setup Processor', function () {
         ).to.be.revertedWith(customError('SetupAlreadyPrepared'));
       });
 
-      it('Retrun correctly the permissions', async () => {
-        const {plugin, helpers, prepareInstallpermissions} =
-          await prepareInstallation(
-            psp,
+      it('retruns correctly the plugin, helpers, and permissions', async () => {
+        const {plugin, helpers, permissions} =
+          await psp.callStatic.prepareInstallation(
             targetDao.address,
             pluginSetupV1Mock.address,
             pluginSetupMockRepoAddress,
@@ -268,7 +265,7 @@ describe('Plugin Setup Processor', function () {
 
         expect(plugin).not.to.be.equal(AddressZero);
         expect(helpers.length).to.be.equal(1);
-        expect(prepareInstallpermissions).to.deep.equal([
+        expect(permissions).to.deep.equal([
           [
             Op.Grant,
             targetDao.address,
@@ -288,7 +285,7 @@ describe('Plugin Setup Processor', function () {
     });
 
     describe('ApplyInstallation', function () {
-      it('Reverts if caller does not have `APPLY_INSTALLATION_PERMISSION`', async () => {
+      it('reverts if caller does not have `APPLY_INSTALLATION_PERMISSION`', async () => {
         // revoke `APPLY_INSTALLATION_PERMISSION_ID` on dao for plugin installer
         // to see that it can't set permissions without it.
         await targetDao.revoke(
@@ -324,7 +321,7 @@ describe('Plugin Setup Processor', function () {
         );
       });
 
-      it("Reverts if PluginSetupProcessor does not have DAO's `ROOT_PERMISSION`", async () => {
+      it("reverts if PluginSetupProcessor does not have DAO's `ROOT_PERMISSION`", async () => {
         // revoke root permission on dao for plugin installer
         // to see that it can't set permissions without it.
         await targetDao.revoke(
@@ -361,7 +358,7 @@ describe('Plugin Setup Processor', function () {
         );
       });
 
-      it('Reverts if plugin setup return the same address', async () => {
+      it('reverts if plugin setup return the same address', async () => {
         const pluginSetupBad = pluginSetupV1MockBad.address;
 
         const dataUser1 = ethers.utils.defaultAbiCoder.encode(
@@ -407,7 +404,7 @@ describe('Plugin Setup Processor', function () {
         ).to.be.revertedWith(customError('SetupAlreadyApplied'));
       });
 
-      it('Correctly complete an instaltion process', async () => {
+      it('correctly complete an instaltion process', async () => {
         const pluginSetup = pluginSetupV1Mock.address;
 
         const {plugin, prepareInstallpermissions} = await prepareInstallation(
@@ -449,7 +446,7 @@ describe('Plugin Setup Processor', function () {
     });
 
     describe('PrepareUninstallation', function () {
-      it('Reverts if `PluginSetupRepo` do not exist on `PluginRepoRegistry`', async () => {
+      it('reverts if `PluginSetupRepo` do not exist on `PluginRepoRegistry`', async () => {
         await expect(
           psp.prepareUninstallation(
             targetDao.address,
@@ -462,7 +459,7 @@ describe('Plugin Setup Processor', function () {
         ).to.be.revertedWith(customError('EmptyPluginRepo'));
       });
 
-      it('Reverts if plugin is not applied yet', async () => {
+      it('reverts if plugin is not applied yet', async () => {
         const pluginSetup = pluginSetupV1Mock.address;
 
         const {plugin, helpers} = await prepareInstallation(
@@ -485,7 +482,7 @@ describe('Plugin Setup Processor', function () {
         ).to.be.revertedWith(customError('SetupNotApplied'));
       });
 
-      it('Reverts if plugin uninstallation is already prepared', async () => {
+      it('reverts if plugin uninstallation is already prepared', async () => {
         const pluginSetupBad = pluginSetupV1MockBad.address;
 
         const installData = ethers.utils.defaultAbiCoder.encode(
@@ -538,7 +535,7 @@ describe('Plugin Setup Processor', function () {
     });
 
     describe('ApplyUninstallation', function () {
-      it('Reverts if caller does not have `APPLY_UNINSTALLATION_PERMISSION`', async () => {
+      it('reverts if caller does not have `APPLY_UNINSTALLATION_PERMISSION`', async () => {
         // revoke `APPLY_INSTALLATION_PERMISSION_ID` on dao for plugin installer
         // to see that it can't set permissions without it.
         await targetDao.revoke(
@@ -565,7 +562,7 @@ describe('Plugin Setup Processor', function () {
         );
       });
 
-      it('Revert if helpers do not match', async () => {
+      it('reverts if helpers do not match', async () => {
         const pluginSetup = pluginSetupV1Mock.address;
 
         const {plugin} = await prepareInstallation(
@@ -587,7 +584,7 @@ describe('Plugin Setup Processor', function () {
         ).to.be.revertedWith(customError('HelpersHashMismatch'));
       });
 
-      it('Revert bad permissions is passed', async () => {
+      it('revert bad permissions is passed', async () => {
         const pluginSetup = pluginSetupV1Mock.address;
 
         const {plugin, helpers, prepareInstallpermissions} =
@@ -626,7 +623,7 @@ describe('Plugin Setup Processor', function () {
         ).to.be.revertedWith(customError('PermissionsHashMismatch'));
       });
 
-      it('Correctly complete an uninstallation process', async () => {
+      it('correctly complete an uninstallation process', async () => {
         const pluginSetup = pluginSetupV1Mock.address;
 
         const {plugin, helpers, prepareInstallpermissions} =
@@ -654,7 +651,7 @@ describe('Plugin Setup Processor', function () {
           EMPTY_DATA
         );
 
-        const event = await decodeEvent(tx, 'UninstallationPrepared');
+        const event = await findEvent(tx, 'UninstallationPrepared');
         const {permissions} = event.args;
 
         await expect(
@@ -687,11 +684,12 @@ describe('Plugin Setup Processor', function () {
     });
 
     describe('PrepareUpdate', function () {
-      it('Reverts if plugin does not support `PluginUUPSUpgradeable` interface', async () => {
+      it('reverts if plugin does not support `IPlugin` interface', async () => {
         const pluginSetupRepoAddr = ADDRESS_TWO;
         const plugin = AddressZero;
-        const pluginUpdateParams = {
+        let pluginUpdateParams = {
           plugin: plugin,
+          oldPluginSetup: pluginSetupV1Mock.address,
           pluginSetupRepo: pluginSetupRepoAddr,
           currentPluginSetup: AddressZero,
           newPluginSetup: AddressZero,
@@ -705,10 +703,31 @@ describe('Plugin Setup Processor', function () {
             helpers,
             EMPTY_DATA
           )
-        ).to.be.revertedWith(customError('PluginNonupgradeable', plugin));
+        ).to.be.revertedWith(customError('IPluginNotSupported', plugin));
       });
 
-      it('Reverts if `PluginSetupRepo` do not exist on `PluginRepoRegistry`', async () => {
+      it('reverts if plugin supports IPlugin, but is non upgradable', async () => {
+        let pluginUpdateParams = {
+          plugin: pluginCloneableMock.address,
+          pluginSetupRepo: ADDRESS_TWO,
+          currentPluginSetup: AddressZero,
+          newPluginSetup: AddressZero,
+        };
+
+        const helpers = [AddressZero];
+        await expect(
+          psp.prepareUpdate(
+            targetDao.address,
+            pluginUpdateParams,
+            helpers,
+            EMPTY_DATA
+          )
+        ).to.be.revertedWith(
+          customError('PluginNonupgradeable', pluginCloneableMock.address)
+        );
+      });
+
+      it('reverts if `PluginSetupRepo` do not exist on `PluginRepoRegistry`', async () => {
         const daoAddress = targetDao.address;
         const pluginSetupV1 = pluginSetupV1Mock.address;
 
@@ -733,7 +752,7 @@ describe('Plugin Setup Processor', function () {
         ).to.be.revertedWith(customError('EmptyPluginRepo'));
       });
 
-      it('Revert if plugin is not applied', async () => {
+      it('revert if plugin is not applied', async () => {
         const daoAddress = targetDao.address;
         const pluginSetupV1 = pluginSetupV1Mock.address;
 
@@ -757,7 +776,7 @@ describe('Plugin Setup Processor', function () {
         ).to.be.revertedWith(customError('SetupNotApplied'));
       });
 
-      it('Revert if helpers passed are missmatched', async () => {
+      it('revert if helpers passed are missmatched', async () => {
         const daoAddress = targetDao.address;
         const pluginSetupV1 = pluginSetupV1Mock.address;
 
@@ -788,7 +807,7 @@ describe('Plugin Setup Processor', function () {
         ).to.be.revertedWith(customError('HelpersHashMismatch'));
       });
 
-      it('Correctly retrun permissions and initData', async () => {
+      it('correctly retruns permissions and initData', async () => {
         const daoAddress = targetDao.address;
         const pluginSetupV1 = pluginSetupV1Mock.address;
 
@@ -829,7 +848,7 @@ describe('Plugin Setup Processor', function () {
         expect(initData).not.to.be.equal('');
       });
 
-      it('Correctly prepare an update', async () => {
+      it('correctly prepares an update', async () => {
         const daoAddress = targetDao.address;
         const pluginSetupV1 = pluginSetupV1Mock.address;
 
@@ -863,7 +882,7 @@ describe('Plugin Setup Processor', function () {
     });
 
     describe('ApplyUpdate', function () {
-      it('Reverts if caller does not have `APPLY_UPDATE_PERMISSION`', async () => {
+      it('reverts if caller does not have `APPLY_UPDATE_PERMISSION`', async () => {
         // revoke `APPLY_INSTALLATION_PERMISSION_ID` on dao for plugin installer
         // to see that it can't set permissions without it.
         await targetDao.revoke(
@@ -890,7 +909,7 @@ describe('Plugin Setup Processor', function () {
         );
       });
 
-      it('Revert if permissions are mismatched', async () => {
+      it('revert if permissions are mismatched', async () => {
         const permissions: any[] = [];
 
         await expect(
@@ -909,7 +928,7 @@ describe('Plugin Setup Processor', function () {
       // it also get threw if UPGRADE_PERMISSION is not granted
       // it('applyUpdate: reverts if PluginNonupgradeable', async () => {});
 
-      it('Correctly process an update', async () => {
+      it('correctly applies an update', async () => {
         const daoAddress = targetDao.address;
         const pluginSetupV1 = pluginSetupV1Mock.address;
 
@@ -941,7 +960,7 @@ describe('Plugin Setup Processor', function () {
           helpers,
           EMPTY_DATA
         );
-        const prepareUpdateEvent = await decodeEvent(
+        const prepareUpdateEvent = await findEvent(
           prepareUpdateTx,
           EVENTS.UpdatePrepared
         );